requirements.txt
Django==2.2.2
django-crispy-forms==1.12.0
djangorestframework=3.14.0
aiohttp==3.8.3
asyncio==3.4.3
redis==4.3.4
celery==5.2.7
